2005-07-23  Stefan Jahn  <stefan@lkcc.org>
        * qucs-lib/qucslib.1: Added manpage for the QucsLib library
        manager program.

        * configure.ac: Finally fixed freebsd[456]* detection.

2005-07-22  Stefan Jahn  <stefan@lkcc.org>

        * configure.ac: Released version 0.0.7 and bumped up to version
        0.0.8.


git-svn-id: https://qucs.svn.sourceforge.net/svnroot/qucs/trunk@658 b5b04e8c-4942-46c9-ab4f-83783d557d1c

+++ qucs/qucsview.cpp
@@ -539,7 +539,7 @@ void QucsView::MovingElements()
       else {  if(long(pw->Port1) & 1) { pw->x1 += MAx1; }
               if(long(pw->Port1) & 2) { pw->y1 += MAy1; } }
 
-      if(int(pw->Port2) > 3) { pw->x2 += MAx1;  pw->y2 += MAy1; }
+      if(((unsigned long)pw->Port2) > 3) { pw->x2 += MAx1;  pw->y2 += MAy1; }
       else {  if(long(pw->Port2) & 1) pw->x2 += MAx1;
               if(long(pw->Port2) & 2) pw->y2 += MAy1; }
 
@@ -725,7 +725,7 @@ void QucsView::MMoveMoving(QMouseEvent *Event)
       else {  if(long(pw->Port1) & 1) { pw->x1 += MAx1; }
               if(long(pw->Port1) & 2) { pw->y1 += MAy1; } }
 
-      if(int(pw->Port2) > 3) { pw->x2 += MAx1;  pw->y2 += MAy1; }
+      if(((unsigned long)pw->Port2) > 3) { pw->x2 += MAx1;  pw->y2 += MAy1; }
       else {  if(long(pw->Port2) & 1) pw->x2 += MAx1;
               if(long(pw->Port2) & 2) pw->y2 += MAy1; }
